+++ src/PhpBrew/Console.php
@@ -5,7 +5,7 @@ use CLIFramework\Application;
 class Console extends Application
 {
     const NAME = 'phpbrew';
-    const VERSION = "1.11.0";
+    const VERSION = "1.11.1";
 
     public function init()
     {