+++ CHANGES.md
@@ -11,6 +11,15 @@ Version 1.11 - Wed Dec  4 13:28:00 2013
 
 - Variant builder is improved with the lookup-prefix
 - Better path detection.
+- Freetype include path fix for +gd variant
+
+    +gd=shared should work for Mac OS platform
+
+- platform libdir is supported, now supports for include/lib paths under 
+
+   $prefix/i386-linux-gnu/
+   $prefix/x86_64-linux-gnu/
+
 
 Version 1.10 - Tue Dec  3 22:55:22 2013